+++ src/analogStuff.cpp
@@ -5,7 +5,7 @@ void initAnalog() {
   initADC(ADC1);
 }
 
-float readVoltage(uint8_t channel) {
+float readVoltage(int channel) {
   //read the voltage from the ADC
 
   //raw ADC is in [0, 1]. Converted to voltage since it takes in 3.3V and outputs 4096 values
@@ -13,6 +13,6 @@ float readVoltage(uint8_t channel) {
   return voltage;
 }
 
-float readUndividedVoltage(uint8_t channel) {
+float readUndividedVoltage(int channel) {
   return readVoltage(channel) * V_DIVIDER;
 }
